feat(training-agent): adopt @adcp/client 5.7 auth helpers
5.7 shipped the remaining SDK fixes we asked for — adopt them and
delete the local workaround files they replace.

- `requireAuthenticatedOrSigned({ signature, fallback, requiredFor,
  resolveOperation })`: the presence-gated composition with required_for
  enforcement. Replaces server/src/training-agent/strict-auth.ts
  (deleted). The SDK's pre-check now fires correctly when the fallback
  throws (bad bearer), closing signed_requests vector 011 which was
  falling through to the required_for path instead of hitting the
  signature-layer malformed-header error.
- `signatureErrorCodeFromCause` + `respondUnauthorized({ signatureError })`:
  SDK unwraps RFC 9421 errors from AuthError.cause and emits
  WWW-Authenticate: Signature challenges. Replaces ~20 lines of
  hand-rolled challenge emission in requireToken.

Also fix: the double-cancel guard in handleUpdateMediaBuy was emitting
INVALID_STATE_TRANSITION, which is a drift code (only appears in 5
prose lines of state-machine.yaml, not in any enum). Canonical seller
code is INVALID_STATE (distinct from the controller-specific
INVALID_TRANSITION). Corrected.

Legacy: 37/55 clean, 282 steps (was 35/55, 279).
Framework: 23/55 clean, 240 steps (was 21/55, 237).
signed_requests: 30/30 passing (was 29/30).
CI floors updated to match.
