+++ .github/workflows/training-agent-storyboards.yml
@@ -38,21 +38,17 @@ jobs:
             flag_value: '0'
             # Baseline on main — keep in sync with the last clean run reported
             # in the PR description. Rising is fine; regressing fails CI.
-            # (37→35 after merging main's two-route design:
-            #  - signed_requests now routes through `/mcp-strict` with
-            #    presence-gated composition; 29/30 vectors pass (vector 011
-            #    malformed-header edge case, follow-up issue).
-            #  - schema_validation past-start rejection reverted to keep
-            #    6 status-derivation unit tests passing; closes as
-            #    accept-and-derive per spec's any_of.)
-            min_clean_storyboards: 35
-            min_passing_steps: 279
+            # 5.7 bump + requireAuthenticatedOrSigned adoption closed
+            # signed_requests vector 011 (malformed-header edge case) and
+            # brought us back above the pre-merge count.
+            min_clean_storyboards: 37
+            min_passing_steps: 282
           - mode: framework
             flag_value: '1'
             # Framework path is opt-in and trails legacy until zod parity
             # lands. Gate keeps it from going backwards while we close the gap.
-            min_clean_storyboards: 21
-            min_passing_steps: 237
+            min_clean_storyboards: 23
+            min_passing_steps: 240
     steps:
       - uses: actions/checkout@v6
